+++ sdk/cosmos/azure-cosmos/CHANGELOG.md
@@ -7,6 +7,7 @@
 #### Breaking Changes
 
 #### Bugs Fixed
+* Fixes not moving the request start time in the `responseStaticsList` - See [PR 35705](https://github.com/Azure/azure-sdk-for-java/pull/35705)
 
 #### Other Changes